Add support to inject other pip index
Prior to this change, python dependencies were only downloaded from the index
defined in requirements.txt. This is a limitation of rules_py, which usually uses
the bazel downloader for python dependencies. However, if a locked
requirements.txt is provided, it directly uses pip to download patches
from the index specified in the file.

Allow specifying custom index URLs through an environment variable.